BoardConfig.mk: set PRESENT_TIME_OFFSET_FROM_VSYNC_NS to 0
As you can see below's portion of log. hwcomposer failed to get display attributes.
on aosp source value of PRESENT_TIME_OFFSET_FROM_VSYNC_NS is 0 in all device trees.
and this thing is related to the hwc + surfaceflinger.So after test I was able to
successfully bypass this issued.

02-23 11:03:18.927   833   833 I hwcomposer: [HWC] getAttributes Display(0)
02-23 11:03:18.927   833   833 I hwcomposer: [HWC] getConfigs Display(1)
02-23 11:03:18.927   833   833 I hwcomposer: [HWC] getAttributes Display(1)
02-23 11:03:18.927   833   833 W hwcomposer: [HWC] Failed to get display attributes (dpy=1 is not connected)
02-23 11:03:18.927   833   833 I hwcomposer: [HWC] getAttributes Display(1)
02-23 11:03:18.927   833   833 W hwcomposer: [HWC] Failed to get display attributes (dpy=1 is not connected)

Signed-off-by: vishalk17 <vishalkapadi95@gmail.com>